pyproject.toml
[project]
name = "data-parsing-py"
version = "0.1.0"
description = ""
authors = [
    {name = "OliverRoat", email = "oliver.roat@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["pyyaml (>=6.0.2,<7.0.0)", "fastapi (>=0.115.8,<0.116.0)", "uvicorn (>=0.34.0,<0.35.0)"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
    { include = "01_Text-based_Data_Formats/01a_Data_Parsing/python" }
]